requirements.txt
requests>=2.28
python-dotenv>=1.0
psycopg2-binary>=2.9
pandas>=1.5